Add some smooth styling to the "game"
Adding some super smooth styling for the etch-a-sketch
as well as for the magnificent button on the top.
This has the purpose to catch the eye of the player
and keep him trapped in an endless loop of playing
the heartbreaking story of etch-a-sketch. Over and over
again until justice has been done.